# openapi.yaml
openapi: 3.0.0
info:
  title: Vehicle Booking System Backend API
  description: Backend service for the Vehicle Booking System.
  version: "1.0"
servers:
  - url: http://localhost:8090
    description: Development server
paths:
  /create-appointment:
    post:
      summary: Create a Rides Booking
      description: Forwards the Booking creation request to the Booking service with proper authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
      responses:
        '200':
          description: Rides Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
  /appointments:
    get:
      summary: Get Rides Booking by email
      description: Fetches Bookings for the given email. Supports filtering by the 'upcoming' query parameter.
      parameters:
        - in: query
          name: email
          schema:
            type: string
            format: email
          required: true
          description: Email address to filter Bookings
        - in: query
          name: upcoming
          schema:
            type: string
            enum: ['true', 'false']
          required: false
          description: Optional filter for fetching only upcoming Bookings
      responses:
        '200':
          description: A list of Bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: Missing or invalid parameters
        '401':
          description: Authentication failed
        '500':
          description: Internal server error
components:
  schemas:
    AppointmentRequest:
      type: object
      properties:
        name:
          type: string
          example: Lasantha Karunarathne
        service:
          type: string
          example: Road Pickup
        phoneNumber:
          type: string
          example: '+9434567890'
        appointmentDate:
          type: string
          format: date-time
          example: '2023-04-30T14:00:00Z'
        email:
          type: string
          format: email
          example: lasa.k@example.com
      required:
        - name
        - service
        - phoneNumber
        - appointmentDate
        - email
    AppointmentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Lasantha Karunarathne
        service:
          type: string
          example: Road Pickup
        phoneNumber:
          type: string
          example: '+9434567890'
        appointmentDate:
          type: string
          format: date-time
          example: '2023-04-30T14:00:00Z'
        email:
          type: string
          format: email
          example: Lasa.ke@example.com
        status:
          type: string
          example: Confirmed
      required:
        - id
        - name
        - service
        - phoneNumber
        - appointmentDate
        - email
        - status